Done Static pages and part of superpals section
----------
 MAIN
 * Finished all the nutrition static pages.
 * Finsihed the My superpals section and the workout requests part.

----------
 NOTE
 *

+++ lib/helpers/api_service.dart
@@ -1,9 +1,11 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:get/get.dart' as GET;
 import 'package:superpal/helpers/constants.dart';
 
 class ApiService {
   late Dio _dio;
+  bool _isRefreshing = false;
 
   ApiService() {
     _dio = Dio();
@@ -24,26 +26,29 @@ class ApiService {
         return handler.next(options);
       },
       onError: (DioError e, ErrorInterceptorHandler handler) async {
-        if (e.response != null &&
+        if (!_isRefreshing &&
+            e.response != null &&
             (e.response!.statusCode == 403 || e.response!.statusCode == 401)) {
+          _isRefreshing = true;
           if (await _refreshToken()) {
             RequestOptions retryOptions = e.requestOptions;
             const FlutterSecureStorage storage = FlutterSecureStorage();
             String? newAccessToken = await storage.read(key: 'access');
-            retryOptions.headers['Authorization'] = 'Bearer $newAccessToken';
+            if (newAccessToken != null) {
+              retryOptions.headers['Authorization'] = 'Bearer $newAccessToken';
+            }
 
             Options retryDioOptions = Options(
               method: retryOptions.method,
               headers: retryOptions.headers,
             );
-
-            return handler.next((await _dio.request(retryOptions.path,
-                options: retryDioOptions,
-                data: retryOptions.data)) as DioException);
+            await _dio.request(retryOptions.path,
+                options: retryDioOptions, data: retryOptions.data);
           } else {
             print('Token refresh failed. Logging out...');
             const FlutterSecureStorage storage = FlutterSecureStorage();
-            storage.deleteAll();
+            await storage.deleteAll();
+            GET.Get.toNamed('/welcome');
           }
         }
 
@@ -69,6 +74,8 @@ class ApiService {
               key: 'refresh', value: refreshedTokens['refresh']);
           return true;
         }
+      } else {
+        GET.Get.toNamed('/welcome');
       }
     } catch (error) {
       print('Token refresh error: $error');
@@ -84,17 +91,90 @@ class ApiService {
     return dio.get('${AppConstants.apiUrl}/superpals/customer/$customerID/');
   }
 
-  Future<Response> getUserSuperpalsWorkoutRequests(String customerID) {
-    return dio.get(
-        '${AppConstants.apiUrl}/superpals/workout_request/customer/$customerID/');
+  Future<Response> getProgress(String customerID) {
+    return dio.get('${AppConstants.apiUrl}/customers/progress/$customerID/');
   }
 
-  Future<Response> getFilteredWorkouts(String intensity, String difficulty) {
+  Future<Response> getUserSuperpalsWorkoutRequests(
+    String customerID,
+    String status,
+  ) {
+    return dio.get(
+        '${AppConstants.apiUrl}/superpals/workout_request/customer/$customerID/',
+        queryParameters: {'status': status});
+  }
+
+  Future<Response> updateUserSuperpalsWorkoutRequests(
+    String request_id,
+    String status,
+  ) {
+    return dio.put(
+        '${AppConstants.apiUrl}/superpals/workout_request/$request_id/',
+        data: {'status': status});
+  }
+
+  Future<Response> getFilteredWorkouts(
+    String intensity,
+    String difficulty,
+  ) {
     return dio.get('${AppConstants.apiUrl}/workouts/',
         queryParameters: {'intensity': intensity, 'difficulty': difficulty});
   }
 
+  Future<Response> postProgress(
+    String customerID,
+    String progressImage,
+  ) {
+    return dio.post(
+      '${AppConstants.apiUrl}/customers/progress/',
+      data: {
+        'user': customerID,
+        'progress_image': progressImage,
+      },
+    );
+  }
+
+  Future<Response> postWorkoutRequest(
+    String customerID,
+    String superPalID,
+    String workoutTime,
+  ) {
+    return dio.post(
+      '${AppConstants.apiUrl}/superpals/workout_request/',
+      data: {
+        'recipient_id': superPalID,
+        'sender_id': customerID,
+        'workout_time': workoutTime,
+      },
+    );
+  }
+
   Future<Response> getCustomersSortedByWorkoutStreak() {
     return dio.get('${AppConstants.apiUrl}/customers/');
+  }
+
+  Future<Response> updateCustomer(
+    String customerID, {
+    String? intensity,
+    String? difficulty,
+  }) async {
+    try {
+      final Map<String, dynamic> requestData = {
+        if (intensity != null) 'workout_intensity': intensity,
+        if (difficulty != null) 'workout_difficulty': difficulty,
+      };
+
+      final response = await _dio.put(
+        '${AppConstants.apiUrl}/customers/$customerID/',
+        data: requestData,
+        options: Options(
+          headers: {'Content-Type': 'application/json'},
+        ),
+      );
+
+      return response;
+    } catch (error) {
+      throw Exception('Error updating customer: $error');
+    }
   }
 }
